requirements.txt
google-generativeai==0.3.2
pylint==3.0.3
pytest==7.4.4
pytest-json-report==1.5.0
python-dotenv==1.0.1
pandas==2.2.0
colorama==0.4.6